requirements.txt
fastapi
uvicorn
pytesseract
Pillow
pdf2image
python-multipart
pymongo
scikit-learn
numpy